requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.7
config==0.5.1
content==0.0.1
Django==4.2.4
django-extensions==3.2.3
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
flake8==6.1.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
phonenumbers==8.13.20
Pillow==10.0.0
platformdirs==3.10.0
pycodestyle==2.11.0
pydot==1.4.2
pyflakes==3.1.0
pyparsing==3.1.1
sqlparse==0.4.4
